Restructured the player logic
Moved the player manager into the base level package with the chunk manager.
Replaced the player package with a module
