--- amulet/api/player/__init__.py
@@ -1,1 +1,0 @@
-from __future__ import annotations
